Add support for mixed content to Documentation
Add attribute mixed of type EFeatureMapEntry[0..*] to support
mixed content as allowed by the XSD.

The text attribute is marked as derived and volatile: it is computed
from the mixed content. The mixed attribute will be saved only
during pure XML serialization and the text attribute only during XMI
serialization (both are not transient).