Remember originally abstract classes
A class that is abstract in the specification, but has to be 
instantiated for proxies, can be made non-abstract and marked as 
originally abstract in the ExtendedMetadata (new Details entry:
abstract -> true, see EventDefinition).

The templates for Factory and ItemProvider are modified to hide 
these classes from the client. They do not appear in the interface 
of Bpmn2Factory. They can not be created in the editor, because
they are not added to the child descriptors in ItemProvider.