Define content types for BPMN 2.0 files
Add BPMN 2.0 XML and XMI content types, derived from XML and 
XMI, respectively. The namespace of the root element is used for
detection, not the extension of the file. This allows arbitrary 
extensions. (Should this prove to be slow the detection could be
narrowed to certain extensions.)
Add content parsers to register the appropriate resource factories
with the content types.

The extensions bpmn2 and bpmn2xmi are explicitly associated with
the content types; otherwise they do not recognize files with these
extensions (conflicts with the extension parser?). 
The extension parsers cannot be removed, because they are 
needed to create new files with the correct resource factory.

Open the default editor also on XMI files.