Correctly resolve import locations on load and save
Interpret the import location as relative to the importing document.
So far, it would be interpreted as relative to the current working
directory, if the resource uri was relative.

BpmnXmlHelper.getCanonicalURI is responsible for resolving relative
URIs to their absolute, canonical form and is used by getPathForPrefix
(on load) and getNsPrefix (on save).

With this commit, all previously failing tests succeed.

Bug: 330017