Test serialization of FormalExpression.body
Test if the body of formal expressions is correctly serialized and
deserialized in XMI and XML. Test the produced XML for XSD-, resp.
XMI-validity.

The current implementation does not support a string object for the
body (only EObjects). However, this would be desirable as the typical
application of formal expression and will be implemented by subsequent
commits.

Bug: 323168