Serialize FormalExpression.body according to XSD
The implementation is similar to the one for Documentation.text. The
body attribute is computed from the mixed property and serialized only
in XMI, while the mixed content itself is serialized when saving as
pure XML.

body was changed from an EAttribute of type EObject to an EReference
of type EString, leading to a XMI serialization as XML attribute,
which is not strictly valid. Thus, the corresponding test still fails.

Bug: 323168